HTML Frame and First game code
A short HTML frame that runs the script was made. The majority of the Javascript was written for the game. It now allows for a player to input a value into the console and have the computer generate a value to play. After collecting the choices, a function returns an enumerated type showing the winner.